Pulls out AttrValue parsing code into its own file.
